Reset and reinitialize MCB if we see it has itself been reset (#103)
Have re-tested today since it has been a while since review start.   Hit MCB reset switch and we re-init from host params.
It has been intentially set to an infrequent check because it is likely a very rare event and I wanted to keep I2C traffic to a minimum.